requirements.txt
Django==4.1.7
cx_Oracle==8.3.0
django-crispy-forms
crispy-bootstrap5
python-decouple==3.7
django-rest-framework
pillow
